+++ scripts/crit-text-config.js
@@ -79,10 +79,13 @@ export class CritTextConfig extends HandlebarsApplicationMixin(ApplicationV2) {
                 "Signika": "Signika"
             },
             fontSizes: {
-                small: "Small",
+                medium: "Medium",
                 normal: "Normal",
                 large: "Large",
-                "extra-large": "Extra Large"
+                "extra-large": "Extra Large",
+                huge: "Huge",
+                massive: "Massive",
+                giant: "Giant"
             },
             letterSpacings: {
                 "tight": "Tight",