+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 0.1.2
 - Decoupled Art Size from Text Size configuration to prevent art scaling dependency on font size.
 - More font sizes
+- More image sizes
 
 # 0.1.1
 - Bug fix: Preview text was triggering a disabled sound