+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Decoupled Art Size from Text Size configuration to prevent art scaling dependency on font size.
 - More font sizes
 - More image sizes
+- Art supports Video
 
 # 0.1.1
 - Bug fix: Preview text was triggering a disabled sound