+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - More image sizes
 - Art supports Video
 - Text template css improv
+- Fill mode Box fixed
 
 # 0.1.1
 - Bug fix: Preview text was triggering a disabled sound