+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Art supports Video
 - Text template css improv
 - Fill mode Box fixed
+- Fill mode Full Screen fixed
 
 # 0.1.1
 - Bug fix: Preview text was triggering a disabled sound